+++ jupyter_kernel_client/__init__.py
@@ -4,8 +4,18 @@
 
 """Jupyter Kernel Client through websocket."""
 
-from jupyter_kernel_client._version import __version__
+from jupyter_kernel_client.__version__ import __version__
+from jupyter_kernel_client.browser_bridge import (
+    BrowserBridgeError,
+    BrowserBridgeServer,
+    BrowserBridgeTimeout,
+    ColabBridge,
+    ColabConnectionInfo,
+    request_colab_connection,
+    request_payload,
+)
 from jupyter_kernel_client.client import KernelClient
+from jupyter_kernel_client.colab import ColabKernelClient
 from jupyter_kernel_client.konsoleapp import KonsoleApp
 from jupyter_kernel_client.manager import KernelHttpManager
 from jupyter_kernel_client.models import VariableDescription
@@ -16,6 +26,12 @@ from jupyter_kernel_client.wsclient import JupyterSubprotocol
 
 __all__ = [
     "SNIPPETS_REGISTRY",
+    "BrowserBridgeError",
+    "BrowserBridgeServer",
+    "BrowserBridgeTimeout",
+    "ColabBridge",
+    "ColabConnectionInfo",
+    "ColabKernelClient",
     "KernelClient",
     "KernelHttpManager",
     "KernelWebSocketClient",
@@ -24,5 +40,7 @@ __all__ = [
     "VariableDescription",
     "JupyterSubprotocol",
     "get_mimebundle_text",
+    "request_colab_connection",
+    "request_payload",
     "__version__",
 ]